+++ src/media/visual/image/raster/macros.rs
@@ -1,4 +1,4 @@
 // devela::media::visual::image::raster::macros
 //
-//!
+//! Defines the [`raster!`] generator.
 //